refactor: Address review feedback
